Update and add tests for CodeQL 2.13+ files.
The tests were generated based on other FOOReaderTest tests - checking
for XSS/SQLi CWE mappings against a minimal file.

This is mostly checking a manually generated file for XSS/SQLi CWE
mappings. This test doesn't actually correspond to the examples in
the source files however, though the SARIF generated is from the
actual CodeQL output, but minified for the purposes of inclusion.

If desired we could include an entire representative output in this,
though we should be concious of it's size - CodeQL includes a large
amount of code flow context in the output.