Update changes to CodeQL reader to be more conservative
This change moves back to checking the rules array on the tool, and
adding checking for the extension. Upon further testing, it appears that
both are equally likely, depending on codeQL usage scenario, and how the
queries/rules are consumed by the CLI. I've added tests for both, I
actually think this is a more conservative change that should avoid
breaking any backcompat.