Update CodeQLReader.java to handle current CodeQL CLI Output. (#45)
* Update CodeQLReader.java to handle current CodeQL CLI Output.

This involves moving rules detection from the rules array that
is on the tool.driver object to parsing the rules run from executed
CodeQL query packs in the tool.extensions Array.

* Fix Potential NullReference Error

* Update and add tests for CodeQL 2.13+ files.

The tests were generated based on other FOOReaderTest tests - checking
for XSS/SQLi CWE mappings against a minimal file.

This is mostly checking a manually generated file for XSS/SQLi CWE
mappings. This test doesn't actually correspond to the examples in
the source files however, though the SARIF generated is from the
actual CodeQL output, but minified for the purposes of inclusion.

If desired we could include an entire representative output in this,
though we should be concious of it's size - CodeQL includes a large
amount of code flow context in the output.

* Update changes to CodeQL reader to be more conservative

This change moves back to checking the rules array on the tool, and
adding checking for the extension. Upon further testing, it appears that
both are equally likely, depending on codeQL usage scenario, and how the
queries/rules are consumed by the CLI. I've added tests for both, I
actually think this is a more conservative change that should avoid
breaking any backcompat.